+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## codem-ffmpeg 0.1.2 (2016/10/21) ##
+
+* Add output property
+
 ## codem-ffmpeg 0.1.1 (2016/09/30) ##
 
 * Added support for cancelling